pyproject.toml
[project]
name = "urban-inspection-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.0",
    "dashscope>=1.25.1",
    "fastapi>=0.121.2",
    "google-genai>=1.50.1",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "numpy>=2.3.4",
    "openai>=2.8.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pycparser==2.21",
    "pydantic>=2.12.4",
    "pymupdf>=1.23.0",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "tenacity>=8.2.0",
    "tokenizers==0.21",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "transformers==4.51.3",
    "uvicorn[standard]>=0.38.0",
    "volcengine-python-sdk[ark]>=4.0.34",
    "werkzeug>=3.1.3",
]
[tool.uv.pip]
extra-index-url = ["https://download.pytorch.org/whl/cu128"]
